Add struct _tracemalloc_runtime_state.
